Namespaced health_checks task and only report failures - useful for crontasks.
